+++ src/networkx_temporal/generators/sbm/dynamic_sbm.py
@@ -1,4 +1,5 @@
 from typing import List, Optional, Union
+from warnings import warn
 
 import networkx as nx
 import numpy as np